Fix CESM tests
