+++ tests/cesm/cesm1-lens-netcdf.json
@@ -1,0 +1,33 @@
+{
+    "esmcat_version": "0.1.0",
+    "id": "sample-cesm1-lens",
+    "description": "This is a sample ESM collection for CESM1-LENS data in netcdf format",
+    "catalog_file": "./tests/cesm/cesm1-lens-netcdf.csv",
+    "attributes": [
+      {
+        "column_name": "experiment",
+        "vocabulary": ""
+      },
+      {
+        "column_name": "case",
+        "vocabulary": ""
+      },
+      {
+        "column_name": "component",
+        "vocabulary": ""
+      },
+      {
+        "column_name": "stream",
+        "vocabulary": ""
+      },
+      { "column_name": "variable", "vocabulary": "" },
+      {
+        "column_name": "member_id",
+        "vocabulary": ""
+      }
+    ],
+    "assets": {
+      "column_name": "path",
+      "format": "netcdf"
+    }
+  }
