+++ tests/cesm/cesm1-lens-zarr.json
@@ -1,0 +1,26 @@
+{
+    "esmcat_version": "0.1.0",
+    "id": "sample-cesm1-lens-aws",
+    "description": "This is a sample ESM collection for CESM1-LENS data in zarr format",
+    "catalog_file": "./tests/cesm/cesm1-lens-aws-zarr.csv",
+    "attributes": [
+      {
+        "column_name": "experiment",
+        "vocabulary": ""
+      },
+      {
+        "column_name": "component",
+        "vocabulary": ""
+      },
+      {
+        "column_name": "frequency",
+        "vocabulary": ""
+      },
+      { "column_name": "variable", "vocabulary": "" }
+
+    ],
+    "assets": {
+      "column_name": "path",
+      "format": "zarr"
+    }
+  }
